component-ize tables
