+++ src/pages/dashboard/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useState, useEffect } from 'react';
-import { useTable, useSortBy } from 'react-table';
 import { LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip, Legend, PieChart, Pie } from 'recharts';
 
 import { sumItemsByDay, groupItemsByCategory } from 'helpers/item';
 import { cloneObject } from 'utils/object';
-import { formatNumberToFloat } from 'utils/currency';
+import { formatNumberToFloat, formatCurrency } from 'utils/currency';
 import { formatDate, getMonthName } from 'utils/date';
 import { UserInterface } from 'components/interfaces/User';
+import ItemTable from 'components/ItemTable';
 
 import styles from 'styles/dashboard.module.scss';
 
@@ -15,64 +15,7 @@ interface DashboardOptions {
   user: UserInterface
 }
 
-function Table({ columns, data }) {
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    rows,
-    prepareRow,
-  } = useTable(
-    {
-      columns,
-      data,
-    },
-    useSortBy
-  )
 
-  return (
-    <>
-      <table {...getTableProps()} className={styles.tableData}>
-        <thead>
-          {headerGroups.map(headerGroup => (
-            <tr {...headerGroup.getHeaderGroupProps()}>
-              {headerGroup.headers.map(column => (
-                // Add the sorting props to control sorting. For this example
-                // we can add them into the header props
-                <th {...column.getHeaderProps(column.getSortByToggleProps())}>
-                  {column.render('Header')}
-                  {/* Add a sort direction indicator */}
-                  <span>
-                    {column.isSorted
-                      ? column.isSortedDesc
-                        ? ' -'
-                        : ' +'
-                      : ''}
-                  </span>
-                </th>
-              ))}
-            </tr>
-          ))}
-        </thead>
-        <tbody {...getTableBodyProps()}>
-          {rows.map(
-            (row, i) => {
-              prepareRow(row);
-              return (
-                <tr {...row.getRowProps()}>
-                  {row.cells.map(cell => {
-                    return (
-                      <td {...cell.getCellProps()}>{cell.render('Cell')}</td>
-                    )
-                  })}
-                </tr>
-              )}
-          )}
-        </tbody>
-      </table>
-    </>
-  )
-}
 
 const DashboardIndex = (props: DashboardOptions) => {
     const [ date, setDate ] = useState('');
@@ -132,32 +75,6 @@ const DashboardIndex = (props: DashboardOptions) => {
       }
     }, [date]);
 
-    const columns = React.useMemo(
-      () => [
-        {
-          Header: 'Data',
-          columns: [
-            {
-              Header: 'Amount',
-              accessor: 'amount',
-            },
-            {
-              Header: 'Category',
-              accessor: 'category',
-            },
-            {
-              Header: 'Date',
-              accessor: 'date',
-            },
-            {
-              Header: 'Type',
-              accessor: 'type',
-            },
-          ],
-        },
-      ],
-      []
-    );
 
     let chartData = sumItemsByDay(data);
     chartData = chartData.map((item) => {
@@ -169,6 +86,7 @@ const DashboardIndex = (props: DashboardOptions) => {
     const tableData = data.map((item) => {
       item = cloneObject(item);
       item.date = formatDate(new Date(item.date));
+      item.amount = formatCurrency(item.amount, 'USD');
       return item;
     });
 
@@ -218,7 +136,7 @@ const DashboardIndex = (props: DashboardOptions) => {
                 </select>
             </label>
         </div>
-        { (date && data.length > 0) ? <div className="col-xs-12 col-sm-6"><Table columns={columns} data={tableData} /></div> : '' }
+        { (date && data.length > 0) ? <div className="col-xs-12 col-sm-6"><ItemTable data={tableData} /></div> : '' }
         
         { chartData.length > 0 ?
             <div className="col-xs-12 col-sm-6">
